+++ Runtime/WebMD5.cs
@@ -1,4 +1,4 @@
-﻿#if UNITY_WEBGL || WEIXINMINIGAME || UNITY_EDITOR
+﻿#if WEIXINMINIGAME && TUANJIE_2022_3_OR_NEWER
 using System;
 using System.Runtime.InteropServices;
 
@@ -6,7 +6,7 @@ namespace WeChatWASM
 {
     public class WebMD5
     {
-       
+
         [DllImport("__Internal", EntryPoint = "_WebMD5")]
         private static extern void getMD5Bytes(IntPtr byteArray, int length, IntPtr result);
 
@@ -27,7 +27,7 @@ namespace WeChatWASM
             Marshal.FreeHGlobal(resultPtr);
 
             return result;
-        } 
+        }
     }
 }
 #endif