+++ Runtime/wechat-default/unity-sdk/texture.js
@@ -121,6 +121,7 @@ const mod = {
             //   err("压缩纹理下载失败！url:"+url);
             mod.reTryRemoteImageFile(path, width, height, limitType);
         };
+        xmlhttp.setRequestHeader('wechatminigame-skipclean', '1');
         xmlhttp.send(null);
     },
     callbackPngFile(path, cid) {
@@ -156,6 +157,7 @@ const mod = {
                 xmlhttp.onerror = function () {
                     mod.reTryRemoteImageFile(path, width, height);
                 };
+                xmlhttp.setRequestHeader('wechatminigame-skipclean', '1');
                 xmlhttp.send(null);
             }
             else {
@@ -236,6 +238,7 @@ GameGlobal.ParalleLDownloadTexture = function (filename) {
                 const p = `${GameGlobal.manager.assetPath}/Textures/${f}/${v.w}/${v.p}.txt`;
                 http.open('GET', p, true);
                 http.responseType = 'arraybuffer';
+                http.setRequestHeader('wechatminigame-skipclean', '1');
                 http.send();
             }
         });